refactor(cli): hoist agent/workspace to cmd/internal/agentworkspace
