+++ cmd/internal/agent/agent.go
@@ -4,8 +4,8 @@ import (
 	"os"
 
 	"github.com/devsy-org/devsy/cmd/flags"
-	"github.com/devsy-org/devsy/cmd/internal/agent/workspace"
 	"github.com/devsy-org/devsy/cmd/internal/agentcontainer"
+	"github.com/devsy-org/devsy/cmd/internal/agentworkspace"
 	"github.com/devsy-org/devsy/pkg/config"
 	"github.com/devsy-org/devsy/pkg/envfile"
 	"github.com/devsy-org/devsy/pkg/log"
@@ -25,7 +25,7 @@ func NewAgentCmd(globalFlags *flags.GlobalFlags) *cobra.Command {
 		Hidden: true,
 	}
 
-	agentCmd.AddCommand(workspace.NewWorkspaceCmd(globalFlags))
+	agentCmd.AddCommand(agentworkspace.NewWorkspaceCmd(globalFlags))
 	agentCmd.AddCommand(agentcontainer.NewContainerCmd(globalFlags))
 	agentCmd.AddCommand(NewDaemonCmd(globalFlags))
 	agentCmd.AddCommand(NewContainerTunnelCmd(globalFlags))
