+++ cmd/internal/agentworkspace/delete.go
@@ -1,4 +1,4 @@
-package workspace
+package agentworkspace
 
 import (
 	"context"
