+++ cmd/internal/agentworkspace/delete_test.go
@@ -1,4 +1,4 @@
-package workspace
+package agentworkspace
 
 import (
 	"os"
