+++ cmd/internal/agentworkspace/docker_darwin_test.go
@@ -1,6 +1,6 @@
 //go:build darwin
 
-package workspace
+package agentworkspace
 
 import (
 	"os"
