+++ cmd/internal/agentworkspace/install_dotfiles.go
@@ -1,4 +1,4 @@
-package workspace
+package agentworkspace
 
 import (
 	"context"
