+++ cmd/internal/agentworkspace/setup_gpg.go
@@ -1,4 +1,4 @@
-package workspace
+package agentworkspace
 
 import (
 	"context"
